Acquire missing lock
There was a semantic merge conflict between the file level conflict
check changes and the log management refactoring. Before the
refactoring, fsHistory was only used in the main transcation processing
thread. The clean up logic was changed in the refactoring, and LSNs
may now be evicted from fsHistory in a client goroutine when they
finish their transaction. This leads to a racy access on the fsHistory
instance.

Acquire the missing mutex before starting a transaction.