operations: Do not resolve revision for zero oid
In most of the operations RPCs we are verifying the expectedOldOID.
And in the documentation of those RPCs it says "If set, it must
either contain a valid, full object ID or the zero object ID in
case the branch should be created. Otherwise, this RPC will return
an error.". But actually these RPCs will return error for the zero
object ID as well because of how git-rev-parse(1) works. It tries
to verify an existing object when ^{object} type peeling operator
appended to the object id. For zero OIDs we should simply skip the
ResolveRevision call.