Record custom hook writes during a transaction
Custom hook modifications have so far been handled in transaction
manager. This was because we didn't expose general methods on a
transaction to stage file system operations as we lacked conflict
checking logic for them. Now we have them, and we can move custom
hook handling out from the transaction manager.

This commit starts recording the custom hook writes as part of
repoutil.SetCustomHooks. We stage the files and directory creations
when the hooks are being modified. Before we exposed general file
system recording methods, we had to provide a hint to transaction
manager to record the hook changes. As the writes can be staged
completely during the transaction now, we can remove all logic
around custom hook handling from transaction manager.

We still keep custom hook related tests in transaction manager.
We'll later trim down the test and generalize them to cover the
various write operations we support rather than the use cases such
as this one. The assertions were changed to reflect the reality of
repoutil.SetCustomHooks leaving an empty directory in the repository
when the hooks are removed. This is now visible in the tests as we
removed the custom hook deletion logic from the tests.