Record repository deletions during transaction
Repository deletion operations have so far been recorded in the
critical section of the transaction manager. This was because we
lacked general file level conflict checks, and couldn't ensure the
operations recorded in the snapshot would still apply in light of
concurrent transactions. This has the downside of pushing the logic
into transaction manager where it doesn't belong. In addition, as
we we're only determining the operations to log in the critical
section, we have to write out the modifications in to the manifest
file and fsync it.

Since we now have file level conflict checks, this commit moves
repository deletion operations to be in the handlers while the
transaction runs. This moves the logic out of the critical section
and allows us to writen and sync the operations into the manifest
file before entering the critical section to take a step closer to
the goal of needing only a single fsync to commit a transaction.

Tests had some adjustments:
- In two tests, the error raised due to a concurrent write has
  changed. This is because we're no longer logging the deletion
  operation in the crtical section using the latest state of the
  repository.
- It's not relevant testing how repository deletions work with other
  modifications anymore as the handler is responsible for recording
  the transaction. It may perform an update first and then delete
  the repository if it wants to.
- We remove the tests covering staging a repository deletion in a
  read-only transaction. The error returned in that case was there
  only to help debugging by receiving a proper error. Deletions would
  already fail if they attempt to remove files or directories from
  a read-only snapshot.