Refactor tests to use RepositoryExists RPC
In order to verify repositories existence, the tests try to access the
repos in the storage directly. This commit refactors the tests to
use RepositoryExists RPC instead. In a follow-up commit, we plan to
acknowledge transactions early after committing them to the WAL.
In this scenario, existence checks via RPC will be useful, as these
calls will automatically apply any pending transactions. We also
refactor other tests who modify the state of the repo to use RPC
instead.