Introduce ForceWALSyncWriteRef workaround
There are still some functions in the tests which directly try access
the repository in the storage and modify the state, this could be a
problem if the pending changes from a different process are not applied
yet to the original repository. So, we introduce a temporary workaround
ForceWALSyncWriteRef which creates a second transaction which has to
wait for the changes to be applied. This would make sure existing
assertions work.