Move recording Mkdir helpers to storage package
The fsrecorder package contains the `storage.FS` implementation
that allows for recording file system operation during a transaction.
The interface is minimal to keep the transcation implementation
itself small. The storage package contains various helpers to record
file system operations using this interface. Move the Mkdir and
MkdirAll helpers to the storage package as they build on the recording
interface but don't need to be part of it.