Implement a no-op storage.FS
When recording with file system operations, we need to work with
relative paths within a storage root. With transactions, we'd use
the `Transaction.FS().Root()` method to acquire the root. If
transactions are not in use, we'd have to pass the storage root
value differently. Implement a no-op FS implmentation that just
returns the correct storage root and doesn't record any operations.
This way we don't need conditionals in code that needs to run with
and without transactions as we can just stub the FS implementation
with a no-op one.