Stage key-value operations outside of critical section
Key-value operations a transaction has performed are currently
recorded into the manifest in the critical section of the transaction
manager. This is unnecessary as all of the operations have already
been done before the criticial section. Stage the performed writes
into the manifest before queuing for commit. This takes us a step
closer to reducing transaction commits to a single fsync as we have
one less write to fsync at commit time.