Reorder critical section
We have some functions that run conditionally depending if a
transaction is targeting a repository. Move the functions together
so we can group them under a single conditional.

File-directory conflict error changes with reftables as reftables
still replay changes in the snapshot. In contrast, files backend
uses our custom conflict checks and return thus a different error
for the same case.