Collect operations under a single conditional
Transactions can target a repository or more generally a partition
if they are writing for example arbitrary key-value state.

Most of the verification operations are specifically for repositories.
We've littered `repositoryTarget` conditionals in them to stub them
out if we're not operating on a repository. Deduplicate the conditional
and collect all these operations under a single conditional.