Record alternate unlinking in handler
We've previously been recording alternate unlink operations in
transaction manager at commit time. This was because we lacked
general file level conflict checks to ensure the operations still
apply if they were recorded in the snapshot while the RPC handler
ran. Now that we have file level conflict checks and recording
functionality, we can start recording the alternate unlink operations
in the handler when they are performed.

We remove the custom alternate unlinking done transaction manager's
tests so far. This highlights a difference in the actual logic and
the tests. Unlinking an alternate from a repository that does not
have one doesn't error but is a no-op. The test is updated.