Link alternate based on relative path in tests
Alternate linking tests are currently taking the content of the
alteranates file. We want to replace the custom alternate linking
logic in the tests with a call to objectpool.Link() and it takes
a relative path instead. Refactor the tests to link using the relative
path of the pool.

One test covering linking multiple alternates was removed. This is not
allowed by the common functionality in the objectpool package and it's
not transaction manager's responsibility to validate the alternate
content.