Record alternate link operations in handler
Alternate linking operation have so far been recorde in transaction
manager as we didn't have a general API for staging file system
operations in transactions. As we have such an API now, move the
linking operation to be recorded in the handler.

We reuse the logic from the handler in the tests and remove the
custom linking logic.