Remove alternate validation logic from transaction manager
Transaction manager should focus on conflict checking concurrent
transactions. Transactions are expected to be correct in isolation
so doing extra validation on them is not necessary. Remove the
validation logic around alternates. We remove tests that were covering
validation failures. Tests that are exercising concurrent behaviors
were updated to assert the conflict.