Demonstrate a bug with deleting a repository
We recently added logic to record reference changes done by a
transaction while it runs. This recording is handled in transaction
manager. Ideally it would live in the handlers but it's not practical
to go and integrate the recording logic in each handler that does
reference updates.

The new logic introduce a bug which prevents deleting repositories
that have a packed-refs file. This is because the packed-refs file
is staged at commit time if it has been modified. Repository deletions
delete the file which looks like a modification and leads to staging
the deletion again. This leads to an invalid operation as the we're
attempting to delete a file in the repository after it has already
been deleted.

Cover this case with a test.