Fix deletion of repositories with packed-refs file
Repository deletions with transaction do not currently work if a
repository contains a packed-refs file. This is because the recently
introduced reference recording logic stages a packed-refs removal if
the file has changed. As a repository removal removes the packed-refs
file, the recorder sees the file has been modified and stages another
removal. Committing the transaction then fails as we attempt to delete
a file from the repository after it has already been removed by the
operations recorded in the handler.

Only stage the packed-refs file changes from the recorder if there
have been reference transactons recorded or if there has been a
housekeeping run that may have modified the file.