Place repository verification operations behind a conditional
Some verification operations really need to only run when the
repository already exists. These are the operations that assert
a repository is in good shape after concurrent writes. Many of
these also require a snapshot so grouping these will help eventually
only creating a snapshot when needed.

These checks need to only run when the repository exists. This is
because repository creations and deletions would already conflict
with every other write. We don't thus need to run these checks if
the repository does not yet exist.

The file system and key-value conflict checks operate on a partition
level and are thus left outside. They also do not need a snapshot.
The logical reference conflict checks are also left outside as they
don't need a repository to function. Transactions would not commit
changes against a non-existent repository though. We don't need to
conflict check reference changes that were introduced as part of
a repository creation as repository creations already conflict with
every operation.