Fsync log entry in the critical section only when needed
Most operations at this point are staged before queuing for commit.
The only exceptions are houskeeping and reftable operations as
these still operate with a staging snapshot and may modify the
manifest or include more files in the log entry.

Write the log entry fully to disk before queuing for commit, and
only flush it to disk again in the critical section if needed.
This improves performance as we don't need to wait do bunch of
fsyncs in the common case when committing a transaction.